+++ plugins/postcss-focus-visible/CHANGELOG.md
@@ -1,4 +1,9 @@
-# Changes to PostCSS Focus Ring
+# Changes to PostCSS Focus Visible
+
+### 2.0.0 (February 17, 2018)
+
+- Changed `:focus-ring` to `:focus-visible` per the specification
+- Removed `assignTo` export option
 
 ### 1.0.0 (May 22, 2017)